@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <RootNamespace>net.jommy.Orleans</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Orleans.Core" Version="3.3.0" />
-    <PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="3.3.0" />
-    <PackageReference Include="Microsoft.Orleans.Runtime.Abstractions" Version="3.3.0" />
+    <PackageReference Include="Microsoft.Orleans.Core" Version="3.6.2" />
+    <PackageReference Include="Microsoft.Orleans.Core.Abstractions" Version="3.6.2" />
+    <PackageReference Include="Microsoft.Orleans.Runtime.Abstractions" Version="3.6.2" />
   </ItemGroup>
 
 </Project>
