@@ -1,10 +1,10 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <PackageVersion>0.2.7</PackageVersion>
-        <AssemblyVersion>0.2.6</AssemblyVersion>
-        <FileVersion>0.2.6</FileVersion>
+        <TargetFramework>net6.0</TargetFramework>
+    <PackageVersion>0.6.0</PackageVersion>
+    <AssemblyVersion>0.6.0</AssemblyVersion>
+    <FileVersion>0.6.0</FileVersion>
     </PropertyGroup>
 
     <ItemGroup>
