@@ -1,17 +1,12 @@
 using System;
 
-namespace net.jommy.RuuviCore.Interfaces
-{
-    [Serializable]
-    public class MeasurementEnvelope
-    {
-        public MeasurementEnvelope()
-        {
-        }
+namespace net.jommy.RuuviCore.Interfaces;
 
-        public string MacAddress { get; set; }
-        public DateTime Timestamp { get; set; }
-        public short SignalStrength { get; set; }
-        public byte[] Data { get; set; }
-    }
+[Serializable]
+public class MeasurementEnvelope
+{
+    public string MacAddress { get; set; }
+    public DateTime Timestamp { get; set; }
+    public short? SignalStrength { get; set; }
+    public byte[] Data { get; set; }
 }
