@@ -31,7 +31,8 @@ namespace net.jommy.RuuviCore.GrainServices
         private readonly DeviceListenerFactory _deviceListenerFactory;
         private readonly ILogger<DBusListener> _logger;
 
-        public DBusListener(IGrainIdentity grainId, Silo silo, ILoggerFactory loggerFactory, IGrainFactory grainFactory, IOptions<DBusSettings> dbusOptions) : base(grainId, silo, loggerFactory)
+        public DBusListener(IGrainIdentity grainId, Silo silo, ILoggerFactory loggerFactory, IGrainFactory grainFactory, IOptions<DBusSettings> dbusOptions) 
+            : base(grainId, silo, loggerFactory)
         {
             _dbusSettings = dbusOptions.Value;
             _deviceListenerFactory = new DeviceListenerFactory(grainFactory, loggerFactory);
@@ -91,8 +92,8 @@ namespace net.jommy.RuuviCore.GrainServices
             }
             catch (Exception e)
             {
-                _logger.LogWarning(e, "DBUS error: {errorMessage}", e.Message);
-                _logger.LogInformation("Ruuvi DBUS Listener NOT listening for bluetooth events.");
+                _logger.LogError(e, "DBUS error: {errorMessage}", e.Message);
+                _logger.LogError("Ruuvi DBUS Listener NOT listening for bluetooth events.");
             }
         }
 
@@ -111,7 +112,7 @@ namespace net.jommy.RuuviCore.GrainServices
             }
             catch (Exception e)
             {
-                _logger.LogWarning(e, "Error occurred when discovering a device: {errorMessage}.", e.Message);
+                _logger.LogError(e, "Error occurred when discovering a device: {errorMessage}.", e.Message);
             }
         }
 
@@ -138,21 +139,31 @@ namespace net.jommy.RuuviCore.GrainServices
             if (manufacturerData != null)
             {
                 var address = await device.GetAddressAsync();
-                // It seems sometimes devices are found again, so let's dispose old observers so DBUS limits will not get exceeded.
-                if (_deviceListeners.Remove(address, out var oldListener))
-                {
-                    _logger.LogDebug("Disposing old observer.", address);
-                    oldListener.Dispose();
-                }
 
+                if (_deviceListeners.TryGetValue(address, out var existingListener))
+                {
+                    if (existingListener.IsAlive())
+                    {
+                        _logger.LogDebug("Using old device listener with address {address} to handle manufacturer data.", address);
+                        await existingListener.HandleDataAsync(manufacturerData);
+                        return;
+                    }
+
+                    // Devices are found again with certain interval. If old listener hasn't had data for a while, let's dispose it and start a new one.
+                    _logger.LogInformation("Disposing old device listener with address {address}.", address);
+                    _deviceListeners.Remove(address);
+                    existingListener.Dispose();
+                }
+                
                 if (_deviceListenerFactory.TryConstructDeviceListener(device, address, manufacturerData, out var deviceListener))
                 {
                     _deviceListeners[address] = deviceListener;
                     await deviceListener.StartListeningAsync();
+                    await deviceListener.HandleDataAsync(manufacturerData);
                 }
                 else
                 {
-                    _logger.LogDebug("Unsupported manufacturer, ignoring.");
+                    _logger.LogInformation("Unsupported manufacturer, ignoring: {data}.", manufacturerData);
                 }
             }
             else
